move _symetric_limits_1d and symetric_limits_2d
