+++ alphadia/peakgroup/utils.py
@@ -197,3 +197,117 @@ def wrap1(
     for i in range(values.shape[0]):
         values[i] = wrap0(values[i], limit)
     return values
+
+
+@nb.njit(cache=USE_NUMBA_CACHING)
+def _symetric_limits_1d(
+    array_1d,
+    center,
+    f=0.95,
+    center_fraction=0.01,
+    min_size=1,
+    max_size=10,
+):
+    """
+    Find the limits of a symetric peak in a 1D array.
+    Allthough the edge is symetrically extended in both directions, the trailing edge will use the last valid value when it reaches the limits of the array.
+
+    Parameters
+    ----------
+
+    array_1d : np.ndarray
+        1D array of intensities
+
+    center : int
+        Index of the center of the peak
+
+    f : float
+        minimal required decrease in intensity relative to the trailing intensity
+
+    center_fraction : float
+        minimal required intensity relative to the center intensity
+
+    min_size : int
+        minimal distance of the trailing edge of the peak from the center
+
+    max_size : int
+        maximal distance of the trailing edge of the peak from the center
+
+    Returns
+    -------
+
+    np.ndarray, dtype='int32', shape=(2,)
+        Array of containing the start and stop index of the peak.
+        The start index is inclusive, the stop index is exclusive.
+        Providing an empty array will return np.array([center, center], dtype='int32').
+        Providing a center outside the array will return np.array([center, center], dtype='int32').
+
+    """
+
+    if len(array_1d) == 0:
+        return np.array([center, center], dtype="int32")
+
+    if center < 0 or center >= array_1d.shape[0]:
+        return np.array([center, center], dtype="int32")
+
+    center_intensity = array_1d[center]
+    trailing_intensity = center_intensity
+
+    limit = min_size
+
+    for s in range(min_size + 1, max_size):
+        intensity = (
+            array_1d[max(center - s, 0)] + array_1d[min(center + s, len(array_1d) - 1)]
+        ) / 2
+        if intensity < f * trailing_intensity:
+            if intensity > center_intensity * center_fraction:
+                limit = s
+                trailing_intensity = intensity
+            else:
+                break
+
+        else:
+            break
+
+    return np.array(
+        [max(center - limit, 0), min(center + limit + 1, len(array_1d))], dtype="int32"
+    )
+
+
+@nb.njit(cache=USE_NUMBA_CACHING)
+def symetric_limits_2d(
+    a,
+    scan_center,
+    dia_cycle_center,
+    f_mobility=0.95,
+    f_rt=0.95,
+    center_fraction=0.01,
+    min_size_mobility=3,
+    max_size_mobility=20,
+    min_size_rt=1,
+    max_size_rt=10,
+):
+    mobility_lower = max(0, scan_center - min_size_mobility)
+    mobility_upper = min(a.shape[0], scan_center + min_size_mobility)
+    dia_cycle_lower = max(0, dia_cycle_center - min_size_rt)
+    dia_cycle_upper = min(a.shape[1], dia_cycle_center + min_size_rt)
+
+    mobility_limits = _symetric_limits_1d(
+        a[:, dia_cycle_lower:dia_cycle_upper].sum(axis=1),
+        scan_center,
+        f=f_mobility,
+        center_fraction=center_fraction,
+        min_size=min_size_mobility,
+        max_size=max_size_mobility,
+    )
+
+    dia_cycle_limits = _symetric_limits_1d(
+        a[mobility_lower:mobility_upper, :].sum(axis=0),
+        dia_cycle_center,
+        f=f_rt,
+        center_fraction=center_fraction,
+        min_size=min_size_rt,
+        max_size=max_size_rt,
+    )
+
+    return mobility_limits, dia_cycle_limits
